move user preference component to dedicated folder
